+++ Pi-Books/GraphQL/Mutations/Inputs/AddPublisherInput.cs
@@ -1,4 +1,4 @@
-﻿namespace Pi_Books.GraphQL.Mutations
+﻿namespace Pi_Books.GraphQL.Mutations.Inputs
 {
     public record AddPublisherInput(string Name);
 }
