+++ Pi-Books/GraphQL/Mutations/Payloads/AddPublisherPayload.cs
@@ -1,6 +1,6 @@
 ﻿using Pi_Books.Data.Models;
 
-namespace Pi_Books.GraphQL.Mutations
+namespace Pi_Books.GraphQL.Mutations.Payloads
 {
     public record AddPublisherPayload(Publisher publisher);
 }
